+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fix cc-vim not working on additional instances
 - Clear instance unique caches on instance destroy
 - Close sound context on instance destroy
+- Disallow window fullscreen state change when creating instances
 
 ## [1.1.3] 2026-06-05